Register to SettingsSaved event of another plug-in in separate thread.
Plug-in load order can cause the depending plug-in to load before
the required plug-in. Assuming both plug-ins have been installed we
can then use `LoadPlugIn` in a separate thread initialed from
`OnLoad` to load the plug-in and register after all.

We still load PluginCommunicationTwo at start up, but let it then load
PluginCommunicationOne as part of its own initialisation.